fix: remove old code
