--- src/components/Global/TokenSelector/Components/index.ts
@@ -1,3 +1,0 @@
-export * from './AdvancedButton'
-export * from './SimpleTokenDisplay'
-export * from './TokenDisplay'
